Safeproof misc/test-stubgenc.sh for the value of a variable
codespell complained

	In misc/test-stubgenc.sh line 15:
	rm -rf $STUBGEN_OUTPUT_FOLDER/*
		   ^----------------------^ SC2115 (warning): Use "${var:?}" to ensure this never expands to /* .

	For more information:
	  https://www.shellcheck.net/wiki/SC2115 -- Use "${var:?}" to ensure this nev...

so I followed the recommendation and also ""quoted the value use later in the lines.
In principle that is all not required ATM since the value is hardcoded in the script BUT
happen the value changes, and e.g. variable name gets changed or mystyped without correct
changes to uses -- someone might end up in pain while triggering all those gotchas.  With
these changes code becomes safer and thus preferable imho.

+++ misc/test-stubgenc.sh
@@ -12,8 +12,8 @@ python -m pip install .
 
 # Remove expected stubs and generate new inplace
 STUBGEN_OUTPUT_FOLDER=./test-data/pybind11_mypy_demo/stubgen
-rm -rf $STUBGEN_OUTPUT_FOLDER/*
-stubgen -p pybind11_mypy_demo -o $STUBGEN_OUTPUT_FOLDER
+rm -rf "${STUBGEN_OUTPUT_FOLDER:?}/"*
+stubgen -p pybind11_mypy_demo -o "$STUBGEN_OUTPUT_FOLDER"
 
 # Compare generated stubs to expected ones
-git diff --exit-code $STUBGEN_OUTPUT_FOLDER
+git diff --exit-code "$STUBGEN_OUTPUT_FOLDER"
